+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Migrate to `security/ignorePageTypeRestriction` flag in TCA
 - Migrate and improve TCA configuration
 - Rename `Userfuncs` namespace to `UserFunctions`
+- Migrate to constructor injection and `Connection` class in `SecurityHeadersMiddleware`
 
 ### Removed
 - Remove obsolete TCEforms tag from FlexForms